Move FindingTypePickerDialog to findings driving/api module
